requirements.txt
kiteconnect
PyYAML
schedule
python-dotenv